Adjusted max-height of main component to handle chat display and updated participant-left event listener to delete channel when only 1 user

+++ react-demo-stream-chat/src/components/ChatWidget/ChatWidget.js
@@ -14,19 +14,19 @@ import 'stream-chat-react/dist/css/index.css';
 import '../ChatWidget/ChatWidget.css';
 
 /**
- * We added state for Chat, just like we have for the Call in Call.js 
+ * We added state for Chat, just like we have for the Call in Call.js
  */
 const STATE_IDLE_CHAT = 'IDLE_CHAT';
 const STATE_JOINING_CHAT = 'JOINING_CHAT';
 const STATE_JOINED_CHAT = 'JOINED_CHAT';
-const STATE_LEFT_CHAT = 'LEFT_CHAT'; 
+const STATE_LEFT_CHAT = 'LEFT_CHAT';
 
 export default function ChatWidget(props) {
   const callObject = useContext(CallObjectContext);
   const [channel, setChannel] = useState(null);
   const [chatClient, setChatClient] = useState(null);
   const [chatState, setChatState] = useState(STATE_IDLE_CHAT);
-  // Storing state boolean as variable 
+  // Storing state boolean as variable
   const displayChat = props.display;
 
   /**
@@ -46,8 +46,8 @@ export default function ChatWidget(props) {
         : 'Guest';
       if (event) {
         /**
-         * NOTE: We're keeping in our demo server so that it's easy to fork, clone, and run this demo. 
-         * But, it's much better to replace the below URL with your own server, so that you can use your own API keys and create your own users. 
+         * NOTE: We're keeping in our demo server so that it's easy to fork, clone, and run this demo.
+         * But, it's much better to replace the below URL with your own server, so that you can use your own API keys and create your own users.
          */
         const joinURL = 'https://opaque-incandescent-count.glitch.me/join-chat';
 
@@ -75,10 +75,10 @@ export default function ChatWidget(props) {
                 chatToken
               );
 
-              // Join the channel using setChannel
+              // Create a channel, using setChannel
               const channel = chatClient.channel('messaging', 'team-chat');
-              channel.watch({ presence: true});
               setChannel(channel);
+
               setChatState(STATE_JOINED_CHAT);
             });
         } catch (error) {
@@ -89,8 +89,8 @@ export default function ChatWidget(props) {
     callObject.on('joined-meeting', joinStreamChat);
   }, [callObject]);
 
-   /**
-   * When a participant leaves the meeting, 'left-meeting', disconnect them from Stream Chat 
+  /**
+   * When a local participant leaves the meeting, 'left-meeting', disconnect them from Stream Chat
    *
    */
   useEffect(() => {
@@ -101,7 +101,7 @@ export default function ChatWidget(props) {
     function leaveStreamChat(event) {
       if (event) {
         try {
-          chatClient.disconnect(); 
+          chatClient.disconnect();
           setChatState(STATE_LEFT_CHAT);
         } catch (error) {
           console.log(error);
@@ -112,20 +112,15 @@ export default function ChatWidget(props) {
   }, [callObject, chatClient]);
 
   /**
-   * When the call is over and everyone has left the meeting, delete the channel 
+   * When only one person is left in the call, delete the channel.
    *
    */
   useEffect(() => {
-
     function clearStreamChat(event) {
-      // Only delete the chat history if everybody has left the call 
-      if (Object.keys(callObject.participants()) > 0) {
-        return;
-      }
-
-      if (event) {
+      const participantCount = Object.keys(callObject.participants()).length;
+      if (event && participantCount === 1) {
         try {
-          channel.delete();  
+          channel.delete();
           setChatState(STATE_LEFT_CHAT);
         } catch (error) {
           console.log(error);
@@ -135,8 +130,9 @@ export default function ChatWidget(props) {
     callObject.on('participant-left', clearStreamChat);
   }, [callObject, channel]);
 
-
-  const chatAvailable = [STATE_JOINING_CHAT, STATE_JOINED_CHAT].includes(chatState);
+  const chatAvailable = [STATE_JOINING_CHAT, STATE_JOINED_CHAT].includes(
+    chatState
+  );
 
   return chatClient && displayChat && chatAvailable ? (
     <Chat client={chatClient} theme={'messaging light'}>
